refactor: rename package from gamedirectory to directory in related components
