+++ HMCL/src/main/java/org/jackhuang/hmcl/ui/directory/GameDirectoryPage.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package org.jackhuang.hmcl.ui.gamedirectory;
+package org.jackhuang.hmcl.ui.directory;
 
 import com.jfoenix.controls.JFXButton;
 import com.jfoenix.controls.JFXTextField;
