+++ HMCL/src/main/java/org/jackhuang/hmcl/ui/versions/GameListPage.java
@@ -52,8 +52,8 @@ import org.jackhuang.hmcl.ui.construct.SpinnerPane;
 import org.jackhuang.hmcl.ui.decorator.DecoratorAnimatedPage;
 import org.jackhuang.hmcl.ui.decorator.DecoratorPage;
 import org.jackhuang.hmcl.ui.download.ModpackInstallWizardProvider;
-import org.jackhuang.hmcl.ui.gamedirectory.GameDirectoryListItem;
-import org.jackhuang.hmcl.ui.gamedirectory.GameDirectoryPage;
+import org.jackhuang.hmcl.ui.directory.GameDirectoryListItem;
+import org.jackhuang.hmcl.ui.directory.GameDirectoryPage;
 import org.jackhuang.hmcl.util.FXThread;
 import org.jackhuang.hmcl.util.io.FileUtils;
 import org.jackhuang.hmcl.util.javafx.MappedObservableList;
